FIX: Search for nearby coverage area with radius 0
The radius' value has set to 3000 (3km). It means that we were looking for any stores that
have supporting areas in radius of 3km. The test spec have citted: "busque o PDV mais próximo e que atenda à mesma". It means that we have to return the POS that have supported the coverage area exactly.